requirements.txt
sentencepiece
protobuf
transformers
scikit-learn
pillow
pandas
numpy
tqdm
matplotlib
seaborn
ipywidgets
widgetsnbextension
notebook
langdetect
sentence-transformers
multilingual-clip
open_clip_torch
ftfy
PyYAML
torch
torchvision
huggingface-hub
einops
accelerate
triton
